+++ CHANGELOG.md
@@ -108,3 +108,4 @@ Other breaking changes:
 - the `fromdict()` and `asdict()` methods in `raillabel.format` classes have been replaced with `from_json()` and `to_json` respectively
 - `raillabel.format.Transform` fields have been changed by `pos -> position` and `quad -> quaternion` to make it more explicit
 - `raillabel.format.FrameInterval` fields have been changed by `frame_start -> start` and `frame_end -> end` to make it more concise
+- `raillabel.format.Frame.uid` field has been removed to avoid redundant information